Compute statistics per group of variables

Compute statistics per group of variables

compute_stat_per_group(stat_col, step_var_map, stats_per_var)

Arguments

stat_col

Name of the column of stats_per_var containing the statistics to compute

step_var_map

A list mapping each variable to a list of variables used at the same step

stats_per_var

A data frame containing statistics per variable

Value

A named vector containing for each variable the sum of the statistics for each group of variables