Filter observation list to exclude situations, variables or dates

Filter observation list to exclude situations, variables or dates

filter_obs(obs_list, var_names = NULL, sit_names = NULL,
  dates = NULL, include = FALSE)

Arguments

obs_list
List of observed values to use for parameter estimation A named list (names = situations names) of data.frame containing one column named Date with the dates (Date or POSIXct format) of the different observations and one column per observed variables with either the measured values or NA, if the variable is not observed at the given date.

var_names
(optional, if not given all variables will be kept) Vector containing the names of the variables to include or exclude

sit_names
(optional, if not given all situations will be kept) Vector containing the names of the situations to include or exclude

dates
(optional, if not given all dates will be kept) Vector containing the dates (POSIXct format) to include or exclude

include
(optional, FALSE by default) Flag indicating if the variables / situations / dates listed in inputs must be included (TRUE) or not (FALSE) in the resulting list

Value

obs_list List of filtered observed values (same format as obs_list input argument)

See also

Examples

obs_list <- list(sit1=data.frame(Date=as.POSIXct(c("2009-11-30","2009-12-10")),
                                 var1=c(1.1,1.5),var2=c(NA,2.1)),
                 sit2=data.frame(Date=as.POSIXct(c("2009-11-30","2009-12-5")),
                                 var1=c(1.3,2)))

# Keep only var1
filter_obs(obs_list,var_names=c("var1"),include=TRUE)

#> $sit1
#>         Date var1
#> 1 2009-11-30  1.1
#> 2 2009-12-10  1.5
#> 
#> $sit2
#>         Date var1
#> 3 2009-11-30  1.3
#> 4 2009-12-05  2.0
#> 

# Exclude observations at date "2009-11-30"
filter_obs(obs_list,dates=as.POSIXct(c("2009-11-30")))

#> $sit1
#>         Date var1 var2
#> 1 2009-12-10  1.5  2.1
#> 
#> $sit2
#>         Date var1 var2
#> 2 2009-12-05    2   NA
#>