Create plots of estimated versus initial values of the parameters

Create plots of estimated versus initial values of the parameters

plot_estimVSinit(init_values, est_values, crit, lb, ub, bubble = TRUE)

Arguments

init_values
Data.frame containing initial values of the parameters for each repetition

est_values
Data.frame containing estimated values of the parameters for each repetition

crit
Vector containing the minimum value of the criterion for each repetition of the minimization

lb
Vector containing the lower bounds of the estimated parameters

ub
Vector containing the upper bounds of the estimated parameters

bubble
Logical indicating if bubbles of size proportional to the minimum values of the criterion should be plot (TRUE, default value) or not (FALSE).

Value

A named list containing one plot per parameter

Details

The number of the repetition that leads to the minimal value of the criterion over all repetitions is written in white (if bubble is TRUE) or in red (if bubble is false) while the other ones are written in black.