sample values in given distribution

sample values in given distribution

sample_params(bounds_list, n, seed = NULL)

Arguments

bounds_list

Either a (optionnally named) list containing a vector of upper and lower bounds (ub and lb), or a named list containing for each parameter the list of situations per group (sit_list) and the vector of upper and lower bounds (one value per group) (ub and lb)

n

The number of values to sample for each parameter

seed

The seed of the random generator (optional, default value=NULL)

Value

A vector or data.frame containing the sampled values (nrow=n)

Details

For the moment only works with uniform distributions but will be hopefully soon extended to any distribution. Use genetic LHS (from lhs package)

Examples

# \donttest{
bounds_list <- list(lb = c(0, 1, 2), ub = c(1, 2, 3))
CroptimizR:::sample_params(bounds_list, 5)
#>            [,1]     [,2]     [,3]
#> [1,] 0.82119557 1.610784 2.484390
#> [2,] 0.48589762 1.201226 2.077456
#> [3,] 0.03768106 1.433481 2.948764
#> [4,] 0.70504393 1.013875 2.678737
#> [5,] 0.28441544 1.840368 2.352563
# }