A wrapper for BayesianTools package

A wrapper for BayesianTools package

wrap_BayesianTools(param_names, obs_list, crit_function, model_function,
  model_options = NULL, optim_options = NULL, prior_information)

Arguments

param_names
Name(s) of parameters to estimate (a parameter name must be replicated if several groups of situations for this parameter)

obs_list
List of observed values to use for parameter estimation A named list (names = situations names) of data.frame containing one column named Date with the dates (POSIXct format) of the different observations and one column per observed variables with either the measured values or NA, if the variable is not observed at the given date.

crit_function
Function implementing the criterion to optimize (optional, default=crit_cwss). See here for more details about the list of proposed criterion.

model_function
Crop Model wrapper function to use

model_options
List of options for the Crop Model wrapper (optional, see help of the Crop Model wrapper function used)

optim_options
List of options of the parameter estimation method: nb_rep, the number of repetitions (optional, default=1) xtol_rel, the tolerance for the stopping criterion on relative differences on parameters values between 2 iterations (optional, default=1e-5) maxeval, the maximum number of criterion evaluation (optional, default=500) path_results, the path where to store the results (optional, default=getwd())

prior_information
Prior information on the parameters to estimate. For the moment only uniform distribution are allowed. Either a list containing:

(named) vectors of upper and lower bounds (ub and lb),

init_values, A data.frame containing initial values to test for the parameters (optional, if not provided (or if less values than number of repetitions of the minimization are provided), the (or part of the) initial values will be randomly generated using LHS sampling within parameter bounds. or a named list containing for each parameter the list of situations per group (sit_list), the vector of upper and lower bounds (one value per group) (ub and lb) and the list of initial values per group init_values (data.frame, one column per group).

Value

prints, graphs and a list containing: