main function for parameter estimation

main function for parameter estimation

estim_param(obs_list, crit_function = crit_log_cwss, model_function,
  model_options = NULL, optim_method = "nloptr.simplex", optim_options,
  param_info, transform_obs = NULL, transform_sim = NULL,
  satisfy_par_const = NULL, var_names = NULL)

Arguments

obs_list
List of observed values to use for parameter estimation A named list (names = situations names) of data.frame containing one column named Date with the dates (Date or POSIXct format) of the different observations and one column per observed variables with either the measured values or NA, if the variable is not observed at the given date.

crit_function
Function implementing the criterion to optimize (optional, see default value in the function signature). See here for more details about the list of proposed criterion.

model_function
Crop Model wrapper function to use.

model_options
List of options for the Crop Model wrapper (see help of the Crop Model wrapper function used).

optim_method
Name of the parameter estimation method to use (optional, see default value in the function signature). For the moment, can be "simplex" or "dreamzs". See here for a brief description and references on the available methods.

optim_options
List of options of the parameter estimation method. path_results The path where to store the results (optional, default=getwd()) Click on the links to see the specific options for the simplex and DreamZS methods.

param_info
Information on the parameters to estimate. Either a list containing:

(named) vectors of upper and lower bounds (ub and lb) (-Inf and Inf can be used),

init_values, A data.frame containing initial values to test for the parameters (optional, if not provided, or if less values than number of repetitions of the minimization are provided), the, or part of the, initial values will be randomly generated using LHS sampling within parameter bounds).

or a named list containing for each parameter the list of situations per group (sit_list), the vector of upper and lower bounds (one value per group) (ub and lb) and the list of initial values per group init_values (data.frame, one column per group, optional). (see here for an example)

transform_obs
User function for transforming observations before each criterion evaluation (optional), see details section for more information

transform_sim
User function for transforming simulations before each criterion evaluation (optional), see details section for more information

satisfy_par_const
User function for including constraints on estimated parameters (optional), see details section for more information

var_names
(optional) List of variables for which the wrapper must return results. By default the wrapper is asked to simulate only the observed variables. However, it may be useful to simulate also other variables, typically when transform_sim and/or transform_obs functions are used. Note however that it is active only if the model_function used handles this argument.

Value

prints, graphs and a list containing the results of the parameter estimation, which content depends on the method used, all that saved in the defined in optim_options.path_results

Details

The optional argument transform_obs must be a function with 4 arguments: o model_results: the list of simulated results returned by the mode_wrapper used o obs_list: the list of observations as given to estim_param function o param_values: a named vector containing the current parameters values proposed by the estimation algorithm o model_options: the list of model options as given to estim_param function It must return a list of observations (same format as obs_list argument) that will be used to compute the criterion to optimize.

The optional argument transform_sim must be a function with 4 arguments: o model_results: the list of simulated results returned by the mode_wrapper used o obs_list: the list of observations as given to estim_param function o param_values: a named vector containing the current parameters values proposed by the estimation algorithm o model_options: the list of model options as given to estim_param function It must return a list of simulated results (same format as this returned by the model wrapper used) that will be used to compute the criterion to optimize.

The optional argument satisfy_par_const must be a function with 2 arguments: o param_values: a named vector containing the current parameters values proposed by the estimation algorithm o model_options: the list of model options as given to estim_param function It must return a logical indicating if the parameters values satisfies the constraints (freely defined by the user in the function body) or not.

See also