sample values in given distribution

sample values in given distribution

sample_params(bounds_list, n, seed = NULL)

Arguments

bounds_list
Either a (optionnally named) list containing a vector of upper and lower bounds (ub and lb), or a named list containing for each parameter the list of situations per group (sit_list) and the vector of upper and lower bounds (one value per group) (ub and lb)

n
The number of values to sample for each parameter

seed
The seed of the random generator (optional, default value=NULL)

Value

A vector or data.frame containing the sampled values (nrow=n)

Details

For the moment only works with uniform distributions but will be hopefully soon extended to any distribution. Use genetic LHS (from lhs package)

Examples

bounds_list <- list(lb = c(0, 1, 2), ub = c(1, 2, 3))
CroptimizR:::sample_params(bounds_list, 5)

#>            [,1]     [,2]     [,3]
#> [1,] 0.02772639 1.435557 2.351448
#> [2,] 0.42881324 1.032942 2.784028
#> [3,] 0.78872354 1.232824 2.160478
#> [4,] 0.82525696 1.892682 2.597242
#> [5,] 0.28881037 1.652949 2.877758